requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-http
pygnmi
